requirements.txt
pandas
librosa
transformers
pytorch-lightning==1.7.2
scikit-learn==0.23.2